pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "domestic abuse support chatbot for Ireland"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypdf>=5.7.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.5.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.1.0",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "markupsafe>=2.1.0",
    "beautifulsoup4>=4.13.4",
    "crawl4ai>=0.7.2",
    "ragas>=0.3.0",
    "datasets>=4.0.0",
]
